Extract test to a non-race test
